% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{case_assign}
\alias{case_assign}
\title{Assigns COVID cases based on individual probabilities}
\usage{
case_assign(df, tmp.dir = getwd(), save_output = TRUE)
}
\arguments{
\item{df}{Input list of the model - output of covid_prob function}

\item{tmp.dir}{Directory for saving a csv recording the number
of new cases each day}

\item{save_output}{Logical. Should the number of new cases be saved as output.}
}
\value{
An updated version of the input list with the new cases assigned
}
\description{
Susceptible individuals are assigned COVID through a Bernoulli draw (rbinom)
based on their probability of becoming a COVID case.
}
