% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{create_input}
\alias{create_input}
\title{Formatting data for infection model}
\usage{
create_input(micro_sim_pop, vars = NULL)
}
\arguments{
\item{vars}{Variables to be kept for use in the infection model}

\item{microm_sim_pop}{The output of the spatial interaction model}
}
\value{
A list of data to be used in the infection model
}
\description{
Formatting the output of the spatial interaction model for use in the
infection model and selecting which variables should be included
}
