% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{infection_length}
\alias{infection_length}
\title{Assigns the infection length of new cases}
\usage{
infection_length(
  df,
  presymp_dist = "weibull",
  presymp_mean = 6.4,
  presymp_sd = 2,
  infection_dist = "normal",
  infection_mean = 14,
  infection_sd = 2,
  timestep,
  tmp.dir = getwd(),
  save_output = TRUE
)
}
\arguments{
\item{df}{Input list of the function - output of an ____assign function}

\item{presymp_dist}{The distribution of the length of the presymptomatic stage}

\item{presymp_mean}{The mean length of the presymptomatic stage}

\item{presymp_sd}{The standard deviation of the length of the presymptomatic stage}

\item{infection_dist}{The distribution of the length of the symptomatic stage}

\item{infection_mean}{The mean length of the symptomatic stage}

\item{infection_sd}{The standard deviation of the length of the symptomatic stage}

\item{timestep}{The day counter}

\item{tmp.dir}{Directory for saving output}

\item{save_output}{Logical. Should output be saved.}
}
\value{
An updated version of the input list with the new cases having
infection lengths assigned
}
\description{
Each new case is assigned a number of days an individual is both
presymptomatic and symptomatic for.
}
