% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{normalizer}
\alias{normalizer}
\title{Changing the spread of values to be between two set values}
\usage{
normalizer(x, lower_bound, upper_bound, xmin, xmax)
}
\arguments{
\item{x}{A number or vector of numbers}

\item{lower_bound}{Desired lower_bound of values}

\item{upper_bound}{Desired upper_bound of values}

\item{xmin}{Expected minimum value of x}

\item{xmax}{Expected maximum value of x}
}
\value{
A number or vector of numbers between the lower and upper bounds
}
\description{
For use in the covid_prob function to change the probabilities
to be 0-1 rather than 0.5-1
}
