% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_status_functions.R
\name{rank_assign}
\alias{rank_assign}
\title{Assign COVID cases by ranking individuals current risk}
\usage{
rank_assign(df, daily_case)
}
\arguments{
\item{df}{Input list of the model - output of covid_prob function}

\item{daily_case}{The desired number of cases that should be assigned on this day}
}
\value{
An updated version of the in put list with new rank assigned cases
added
}
\description{
Used to assign cases for the time period in which the model is being seeded.
Individuals are ranked in descending order by current risk and then desired number
of cases are assigned to the highest ranking individuals.
}
\details{
Will be run in place of the case_assign function when the model is being seeded.
}
